+++ src/routes/bulk/mod.rs
@@ -14,21 +14,5 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-pub mod bulk;
-pub mod check_email;
-pub mod version;
-
-use super::errors;
-use sqlx::{Pool, Postgres};
-use warp::Filter;
-
-pub fn create_routes(
-	conn_pool: Pool<Postgres>,
-) -> impl Filter<Extract = impl warp::Reply, Error = warp::Rejection> + Clone {
-	version::get::get_version()
-		.or(check_email::post::post_check_email())
-		.or(bulk::post::create_bulk_email_vrfy_job(conn_pool.clone()))
-		.or(bulk::get::get_job_status(conn_pool.clone()))
-		.or(bulk::get::get_job_result(conn_pool))
-		.recover(errors::handle_rejection)
-}
+pub mod get;
+pub mod post;