+++ src/routes/check_email/mod.rs
@@ -1,5 +1,5 @@
 // Reacher - Email Verification
-// Copyright (C) 2018-2021 Reacher
+// Copyright (C) 2018-2022 Reacher
 
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU Affero General Public License as published
@@ -14,6 +14,4 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-pub mod header;
-mod known_errors;
 pub mod post;